Added datetime to db
A bug where main.auto_placement placed the smaller in such a way that
the largest ships could not be placed caused the function to never
terminate. Should be leviated by placing the largest ships first.

If the module readchar is installed, replays can now be quitted by
pressing q.